% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymeasures.R
\name{alpha_ci}
\alias{alpha_ci}
\title{Cronbach's alpha Confidence Interval}
\usage{
alpha_ci(a, n, p, sig = 0.05, include_a = TRUE)
}
\arguments{
\item{a}{A scalar numeric. The value of alpha to be tested}

\item{n}{A scalar numeric. The number of cases (subject, respondents).}

\item{p}{A scalar numeric. The number of item in the scale.}

\item{sig}{A scalar numeric between 0 and 1. Default 0.05 (95% confidence).}
}
\description{
Calculate a confidence interval for Cronbach's alpha using the F-distribution
}
