% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationtasks.R
\name{ARImpute}
\alias{ARImpute}
\title{ARImputation}
\usage{
ARImpute(cars, dt, var_names, ari_control = arulesimp_control(method =
  "best_rule", use_default_classes = TRUE))
}
\description{
Association Rules-based Imputation
}
