% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationtasks.R
\name{arulesimp_control}
\alias{arulesimp_control}
\title{Control parameters for synth_missing}
\usage{
arulesimp_control(method, support, confidence, lift)
}
\arguments{
\item{method}{Character string. Allowed values are "CBA"}
}
\description{
Control the parameters for generating patterns of missing data
}
