% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationtasks.R
\name{cars_control}
\alias{cars_control}
\title{Control parameters for make_cars}
\usage{
cars_control(support, confidence, lift, sort_by = "confidence")
}
\arguments{
\item{support}{A scalar value between 0 and 1. Minimum support to filter rules.}

\item{confidence}{A scalar value between 0 and 1. Minimum confidence to filter rules.}

\item{lift}{An optional scalar value between 0 and 1. Minimum lift to filter rules.}

\item{sort_by}{A character string (default "confidence") indicating which quality measure to sort the rules by}

\item{ruleset}{An object of class rules (package arules)}

\item{var_names}{A character vector with the names of target variables in the dataset which the rules came from}
}
\description{
Control the parameters for generating patterns of missing data
}
