% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticdata.R
\name{compare_dewinter}
\alias{compare_dewinter}
\title{Compare a Likert item to dewinter distributions}
\usage{
compare_dewinter(v, de_dist = "all")
}
\description{
Compare a Likert item to dewinter distributions
}
