% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticdata.R
\name{compare_plumpton}
\alias{compare_plumpton}
\title{Compare a Likert item to Plumpton distributions}
\usage{
compare_plumpton(v, pl_dist = "all")
}
\description{
Compare a Likert item to Plumpton distributions
}
