% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{cut_equal}
\alias{cut_equal}
\title{Equal cuts to item scale}
\usage{
cut_equal(dt, dist = "equal_distance", levs = 5, labels = FALSE)
}
\description{
COnvenience function for generating discrete data from continuous
}
