% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulaterules.R
\name{extract_cars}
\alias{extract_cars}
\title{Converts rules into a format that is more programmable}
\usage{
extract_cars(cars, v)
}
\description{
Converts rules into a format that is more programmable
}
