% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationtasks.R
\name{find_cols_to_impute}
\alias{find_cols_to_impute}
\title{Find rows to impute}
\usage{
find_cols_to_impute(dt, which_rows)
}
\arguments{
\item{dt}{A data frame or an object of class mim}

\item{which_rows}{A integer vector of row numbers to search.}
}
\description{
Find the columns (variables) with missing values to impute for each target row
}
