% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationtasks.R
\name{find_rows_to_impute}
\alias{find_rows_to_impute}
\title{Find rows to impute}
\usage{
find_rows_to_impute(df, which_cols)
}
\arguments{
\item{df}{A data frame}

\item{which_cols}{A vector of variable names to search.}
}
\description{
Find the rows with missing values to impute for each target variable
}
