% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationtasks.R
\name{iteration_control}
\alias{iteration_control}
\title{Control parameters for AR_iter_Imputation}
\usage{
iteration_control(method, splits, max_iter, target_convergence, class_balance)
}
\arguments{
\item{method}{A character string. Allowed values are "none", "propensity"}

\item{splits}{A scalar numeric denoting the number of propensity groups. If not an integer will be rounded. Ignored if method = "none"}

\item{max_iter}{A scalar numeric. If not an integer will be rounded.}

\item{target_convergence}{A scalar numeric. If not an integer will be rounded.}

\item{class_balance}{No class balancing will be applied if not set. A list of paramaters to be passed to ROSE::ovun.sample.}
}
\description{
Control the parameters for the AR iterative imputation routine
}
