% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationtasks.R
\name{LikertImpute}
\alias{LikertImpute}
\title{Person Mean Imputation}
\usage{
LikertImpute(dt, dt_mim, method = "PM", rounding, top_code = FALSE,
  bottom_code = FALSE)
}
\arguments{
\item{dt}{A data frame containing some missing values}

\item{rounding}{A scalar character, indicating the type of rounding to use}
}
\description{
Impute missing values by substituting the row mean of observed values
}
