% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymeasures.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(theta, theta_p)
}
\arguments{
\item{theta}{A numeric vector containing the parameter to be tested.}

\item{theta_p}{A scalar numeric. The population or reference value}
}
\description{
Calculate the mean absolute error of a parameter with respect to another value, usually the population parameter
}
