% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationtasks.R
\name{make_cars}
\alias{make_cars}
\title{Create classification rules}
\usage{
make_cars(dt, var_names, c_control = cars_control(consequent = NULL, support =
  0.1, confidence = 0.1, sort_by = "confidence"))
}
\description{
Create a list of classification rules for each target viable
}
