% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulaterules.R
\name{make_cars_old}
\alias{make_cars_old}
\title{Creat classification rules}
\usage{
make_cars_old(ruleset, var_names, min_supp = 0.1, min_conf = 0.1,
  min_lift = 1)
}
\arguments{
\item{ruleset}{An object of class rules (package arules)}

\item{var_names}{A character vector with the names of target variables in the dataset which the rules came from}

\item{min_supp}{A scalar value between 0 and 1. Minimum support to filter rules.}

\item{min_conf}{A scalar value between 0 and 1. Minimum confidence to filter rules.}

\item{min_lift}{A scalar value between 0 and 1. Minimum lift to filter rules.}
}
\description{
Create a list of classification rules for each target viable
}
