% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{mimstats}
\alias{mimstats}
\title{Missing Indicator Matrix Statistics}
\usage{
mimstats(dt, one_as_missing = TRUE)
}
\arguments{
\item{dt}{A data frame}

\item{one_as_missing}{A scalar boolean. If true, 1 represents missing and 0 represents observed data. Otherwise the reverse true.}
}
\value{
A list of ...
}
\description{
Counts and means of missing indicator matrix
}
\examples{
mims(matrix(c(1, rep(0, 7), 1), ncol = 3))
}
