% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymeasures.R
\name{mse}
\alias{mse}
\title{Mean Squared Error}
\usage{
mse(theta, theta_p)
}
\arguments{
\item{theta}{A numeric vector containing the parameter to be tested. A scalar value will return Inf}

\item{theta_p}{A scalar numeric.}
}
\description{
Calculate the mean squared error of a parameter with respect to another value, usually the population parameter
}
