% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{no_missing_check}
\alias{no_missing_check}
\title{Check for any missing values}
\usage{
no_missing_check(dt)
}
\description{
This utility function is a simple, routine check for presence of missing data. Used internally, not exported.
}
