% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{ord_combi_expand}
\alias{ord_combi_expand}
\title{Rounded average over Likert Scales}
\usage{
ord_combi_expand(dt, likert_scales, keep_orig = TRUE)
}
\description{
Create a new ordinal variable by averaging and rounding to nearest integer over any combination of ordinal variables.
}
