% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{ord_cum_combine}
\alias{ord_cum_combine}
\title{Ordinal to cumulative binary representation}
\usage{
ord_cum_combine(dt, var_names, mn, mx, stride = 1, keep_orig = FALSE)
}
\description{
Expand ordinal variables to a group of binary variables representing cumulative values
}
