% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{ord_grp_combine}
\alias{ord_grp_combine}
\title{Ordinal to grouped binary representation}
\usage{
ord_grp_combine(dt, var_names, mn, mx, stride = 1, keep_orig = TRUE)
}
\description{
Expand ordinal variables to binary variables representing groups of adjacent values
}
