% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationtasks.R
\name{PMImpute}
\alias{PMImpute}
\title{Person Mean Imputation}
\usage{
PMImpute(dt, rounding = "nd")
}
\arguments{
\item{dt}{A data frame containing some missing values}

\item{rounding}{A scalar character, indicating the type of rounding to use}
}
\description{
Impute missing values by substituting the row mean of observed values
}
