% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymeasures.R
\name{quality_measures}
\alias{quality_measures}
\title{Quality Measures of Parameter Estimate}
\usage{
quality_measures(theta, theta_p)
}
\arguments{
\item{theta}{A numeric scalar or vector containing the parameter to be tested}

\item{theta_p}{A scalar numeric. The population or reference value}
}
\description{
Calculate various quality measures for a parameter with respect to a population or reference value
}
