% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymeasures.R
\name{rel_bias}
\alias{rel_bias}
\title{Relative Bias}
\usage{
rel_bias(theta, theta_p)
}
\arguments{
\item{theta}{A numeric scalar or vector containing the parameter to be tested}

\item{theta_p}{A scalar numeric. The population or reference value}
}
\description{
Calculate the relative bias of a parameter with respect to another value, usually the population parameter
}
