% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationtasks.R
\name{rules_to_cars}
\alias{rules_to_cars}
\title{Convert arules object to cars}
\usage{
rules_to_cars(ruleset, var_names, sort_by)
}
\description{
Get the classification association rules from a rule set for named variables
}
