% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticdata.R
\name{synth_dewinter}
\alias{synth_dewinter}
\title{Synthesise Likert scale data using dewinter distributions}
\usage{
synth_dewinter(var_names, dists, n = 10000)
}
\description{
Synthesise Likert scale data using dewinter distributions
}
