% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticdata.R
\name{synth_plumpton}
\alias{synth_plumpton}
\title{Synthesise Likert scale data using Plumpton distributions}
\usage{
synth_plumpton(var_names, dists, n = 10000)
}
\description{
Synthesise Likert scale data using Plumpton distributions
}
