% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticdata.R
\name{synth_wu_data}
\alias{synth_wu_data}
\title{Synthesise Likert Scale data from latent variable models}
\usage{
synth_wu_data(model, sample_size, seed)
}
\description{
Synthesise Likert Scale data using method from Wu
}
