% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymeasures.R
\name{t_ci}
\alias{t_ci}
\title{Students T based confidence interval}
\usage{
t_ci(stat, st_dev, n, sig = 0.05, include_s = TRUE)
}
\arguments{
\item{n}{A scalar numeric. The number of cases (subject, respondents).}

\item{sig}{A scalar numeric between 0 and 1. Default 0.05 (95% confidence).}

\item{include_s}{A scalar boolean. Whether to include the input statistic in the output.}

\item{a}{A scalar numeric. The value to be tested.}
}
\description{
Convenience function to calculate confidence interval for a statistic and its standard error.
}
