% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{unique_values}
\alias{unique_values}
\title{Create a list of unique values in a dataset}
\usage{
unique_values(dt, which_cols)
}
\arguments{
\item{dt}{A data frame}

\item{which_cols}{A vector of variable names to search.}
}
\value{
A named list. Names are the input variable names.
List values are vectors of unique values from the input dt.
}
\description{
Useful for searching for rules
}
\examples{
missing_values(data.frame(x = c("foo", NA), y = c(NA, "bar")))
}
