% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{all_factor}
\alias{all_factor}
\title{Convert all variables to factors.}
\usage{
all_factor(dt, refactor = FALSE)
}
\arguments{
\item{dt}{A data frame}

\item{refactor}{Logical. Do you want it to also work on variables that are already factors? (Default is FALSE)}
}
\value{
A copy of the input where all variables except logicals have been converted to factors
}
\description{
arules wants only logical or factor. This function converts everything it finds into a factor
}
\examples{
str(all_factor(data.frame(x = c("foo", "bar"), y = c(10, 20))))
}
