% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationtasks.R
\name{arulesimp_control}
\alias{arulesimp_control}
\title{Control parameters for ARImputation}
\usage{
arulesimp_control(method, use_default_classes, rows_to_impute, top_n,
  weighted_chisq)
}
\arguments{
\item{ruleset}{An object of class rules (package arules)}

\item{var_names}{A character vector with the names of target variables in the dataset which the rules came from}

\item{support}{A scalar value between 0 and 1. Minimum support to filter rules.}

\item{confidence}{A scalar value between 0 and 1. Minimum confidence to filter rules.}

\item{lift}{An optional scalar value between 0 and 1. Minimum lift to filter rules.}

\item{sort_by}{A character string (default "confidence") indicating which quality measure to sort the rules by}
}
\description{
Control the parameters for the AR imputation routine
}
