% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{char_to_na}
\alias{char_to_na}
\title{Replace character strings with NA.}
\usage{
char_to_na(dt, char_string = "")
}
\arguments{
\item{char_string}{A character string (default = "")}

\item{df}{A data frame}
}
\value{
A copy of the input with all target strings replaced by NA
}
\description{
Useful for data with empty strings (default) or other character values to indicate missing
}
\examples{
char_to_na(data.frame(x = c("foo", ""), b = c("", "bar")))
char_to_na(data.frame(x = c("foo", "_"), b = c("_", "bar")), char_string = "_")
}
