% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{cut_rhemtulla}
\alias{cut_rhemtulla}
\title{Normal to Item Scale}
\usage{
cut_rhemtulla(dt, dist = "symmetric", levs = 5, labels = FALSE)
}
\description{
Discretize normally distributed data to ordinal scales based on desired distribution and number of levels
}
