% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{missing_matrix}
\alias{missing_matrix}
\title{Create a missing indicator matrix for a dataset}
\usage{
missing_matrix(dt, one_as_missing = TRUE)
}
\arguments{
\item{dt}{A data frame}

\item{one_as_missing}{A scalar boolean. If true, 1 represents missing and 0 represents observed data. Otherwise the reverse true.}
}
\value{
A list containing matrix of 1 and 0 having the same dimensions as the input and Carpita's measures.
}
\description{
Converts a data frame into a missing indicator matrix
}
\examples{
missing_matrix(data.frame(x = c("foo", NA), y = c(NA, "bar")))
}
