% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{missing_values}
\alias{missing_values}
\title{Find the variables that contain any missing (NA) values.}
\usage{
missing_values(dt, sorted = TRUE)
}
\arguments{
\item{dt}{A data frame}

\item{sorted}{A logical vector. Do you want to sort the results (Default = TRUE)}
}
\value{
A named numeric. Names are the variable names from the input dt. Values are the count of NA found in that variable
}
\description{
Find the variables that contain any missing (NA) values.
}
\examples{
missing_values(data.frame(x = c("foo", NA), y = c(NA, "bar")))
}
