% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticdata.R
\name{synth_missing}
\alias{synth_missing}
\title{Synthesise missing data patterns}
\usage{
synth_missing(dt, syn_control = missing_control(pattern = "MCAR"),
  prob = 0.2, exact = TRUE)
}
\description{
Synthesise missing data patterns
}
